feat(infra): Implementing endpoint to delete user by id

+++ src/main/java/com/spring/Ecommerce/core/gateway/UserGateway/UserDeleteGateway.java
@@ -1,5 +1,5 @@
 package com.spring.Ecommerce.core.gateway.UserGateway;
 
 public interface UserDeleteGateway {
-    void execute(int id);
+    void execute(Long id);
 }
