+++ src/main/java/com/spring/Ecommerce/core/useCases/Users/UserDeleteUseCase.java
@@ -1,5 +1,5 @@
 package com.spring.Ecommerce.core.useCases.Users;
 
 public interface UserDeleteUseCase {
-    void execute(int id);
+    void execute(Long id);
 }
